Refactored code, added enumerations, modal validation and corrected responses and updated tests

+++ EagleBank.Orchestration/code/Implementation/UserDetailsOrchestrator.cs
@@ -7,22 +7,28 @@ using EagleBank.Repository.Interfaces;
 
 namespace EagleBank.Orchestration.code.Implementation
 {
-    public class UserDetailsOrchestrator(IUserRepository userRepository) : IUserDetailsOrchestrator
+    public class UserDetailsOrchestrator(IUserRepository userRepository, IBankAccountRepository bankAccountRepository) : IUserDetailsOrchestrator
     {
         readonly IUserRepository _userRepository = userRepository ?? throw new ArgumentNullException(nameof(userRepository));
-
+        readonly IBankAccountRepository _bankAccountRepository= bankAccountRepository ?? throw new ArgumentNullException(nameof(bankAccountRepository));
         public async Task<UserDetailsResponse> AddUserDetailsAsync(UserDetailsRequest request)
         {
-            var dtoRequest = request.Convert(Guid.NewGuid().ToString());
+            var userId = $"usr-{new Random().Next(1, 10000).ToString()}";
+            var dtoRequest = request.Convert(userId);
             dtoRequest.CreatedTimeStamp = DateTime.Now;
             dtoRequest.UpdatedTimeStamp = DateTime.Now;
             var response = await _userRepository.AddUserDetailsAsync(dtoRequest);
             return response.Convert();
         }
 
-        public Task<bool> DeleteUserDetailsAsync(string userId)
+        public async Task<bool> DeleteUserDetailsAsync(string userId)
         {
-            var response = _userRepository.DeleteUserDetailsAsync(userId);
+            var bankAccounts =  await _bankAccountRepository.GetAllBankAccountsAsync(userId);
+            if (bankAccounts != null)
+            {
+                throw new ConflictErrorException("A user cannot be deleted when they are associated with a bank account");
+            }
+            var response = await _userRepository.DeleteUserDetailsAsync(userId);
             return response;
         }
